Rework Interaction and Relations
Now Interaction owns all the data (in the _data boost::array), which are
BlockVector (linked to the real data in the DS). A Relation contains
only the plugins and the different Jacobians. This means that we can use
a singl Relation with multiple Interactions. This should reduce the
memory usage with multiples DS.

3 tests are still failing and maybe more, please check !

On a side note, this commit also introduce some fixes to crosscompile
Numerics to windows

git-svn-id: svn+ssh://gforge/svn/siconos/trunk@3373 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5