Begin port to C++11 mode
We use some function that are now in the standard like
- smart_ptr
- bind
- tie
- array

Currently there are some preprocessor directives to determine if we are
compiling in c++11 mode or not. Right now it works with >=gcc-4.7 and
clang. The ultimate goal is to enable a Siconos user (installing binaries)
to run examples without boost. Plus it is always good to use the
standard lib instead of a 3rd party.

git-svn-id: svn+ssh://gforge/svn/siconos/trunk@3456 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5