[Numerics] spot big bug in cs_aaxpy -> add assert until fixes
I'm a bit overload right now and I'm a little bit unfamilliar with the data layout,
but the code is given wrong result for any value of alpha != 1
The only occurence of this is in globalLineSearchGP, with alpha = 0.

We should add tests to test basis matrix operation.

Feel free to fix this, otherwise I'll try to do ASAP