[kernel, io, control] Replace __cplusplus with SICONOS_CXXVERSION in headers.
The C++ version is used to select between C++11 and Boost shared_ptr
definitions, which changes the ABI.  Therefore it is necessary for
external codebases using the Siconos headers to define the same
interface.  SICONOS_CXXVERSION is a copy of __cplusplus defined at
configure time, and therefore is the one used to compile Siconos.

Without this patch, C++11 code using Siconos compiled with C++98 will
result in a successful compilation, but mysteriously fail to link.