[cmake] Add detection of C++1x features, and store results in SiconosConfig.h.
The use of boost::shared_ptr or std::shared_ptr is detected by the
preprocessor by checking against __cplusplus in some headers. So
software including those headers when -std=c++11 was enabled would
incorrectly choose the std:: namespace, when Siconos was compiled with
the boost:: namespace. This caused linker errors.

This patch:

  * Adds detection of individual C++11 features, such as
    std::shared_ptr, std::array, std::unordered_map, std::to_string,
    std::bind.

  * Detection variables are stored in SiconosConfig.h
    e.g., SICONOS_STD_SHARED_PTR, etc..

  * Detection of individual features is only performed if CXXVERSION
    or CMAKE_CXX_STANDARD is changed.

  * User flags to turn off use of these features are also stored in
    SiconosConfig.h.  They default to ON to keep compatibility with
    previous behaviour.

    - SICONOS_USE_BOOST_FOR_CXX11
    - SICONOS_USE_MAP_FOR_HASH

  * Fixes some problematic areas in mechanics where boost::array or
    similar was used even when c++11 was enabled.

  * Replaces checks against __cplusplus in the headers with checks
    against specific feature variables.

Note that CMake's own mechanism for setting the C++ standard must be
used to select C++11 or C++14:

    cmake .. -DCMAKE_CXX_STANDARD=11

This option is explicitly passed on to the try_run() calls used for
feature detection.