[numerics/py] add newobjects instructions for SWIG
When a C function allocates a struct and return a pointer to it, one of the
calling function in the stack is usually in charge of deleting the
memory. This does not work out of the box with the swig bindings. The
right way to do so is to do this is to mark the function as returning a
pointer to newly allocated memory that swig has to deallocate. For
instance globalFrictionContact_fclib_read return a
GlobalFrictionContactProblem* that swig has to deallocate. Therefore we
mark it as such. The list in numerics_newobjects.i may not be
comprehensive. Feel free to complete it.