[kernel/py] rework graph support because of memory issue
when a list of dynamicalsystem was created from a siconosgraph, its type
was a wrapper around a std::vector. However there where some memory
management issue: when an element was requested from the list, the
refcount on the SP was not properly updated and we would use a memory
bank that has been freed.

This rework the graph mamangement by returning a tuple of DS or
Interaction instead of an std::vector. Plus now we perform a conversion
to return the true DS type and not a DynamicalSystem