[numerics] fix HUGE BUG in gfc3d_compute_error
Before this commit, the error was set to
\sqrt{ \|error_velocity\|_2 + (r - \proj_K{ r - u })^2 }

roughly \sqrt{ \|error_velocity\|_2 + F_nat^T F_nat }

The double square root made it almost impossible for any algorithm to
meet a good convergence criterion.

This bug was uncovered by 16546659cd9659ba8ab0e66d71217985114c461b
which added the square root.

If the error should be computed with a different formula, we just have
to agree on which one.