[mechanics proposed] Skeleton of a new approach to the Mechanics API.
User code describes a scene consisting of BodyDS (::NewtonEulerDS) and
contactor surfaces, using objects not connected to a specific
implementation.

The broadphase implementation (e.g. Bullet) visits this scene and
composes its own mirrored scene graph.  Thus, the implementation is
isolated from the scene description, and the contactor shapes are
isolated from the DS, as opposed to now, where we have Bullet-specific
objects and shape-specific DSs.

A callback mechanism is used to keep the implementation up to date
when object properties change.  The current implementation here of a
an onChanged() hook is very tentative, but it demonstrates using it to
put the objects on a "dirty" list.

This patch doesn't completely implement this idea, but some pieces are
in place.  Notably missing: many shapes (plane, disc, mesh...), a
functioning buildInteractions(), support for joints.  It is planned to
borrow quite a bit from SpaceFilter and BulletSpaceFilter.