[numerics/py] compile python module in C
for no obvious reason the python module was compiled in C++.
This commit make it compile in C, unless WITH_CXX is specified
Most remove code was dead. The only loss if that __str__() on
NumericsMatrix does not work now. However, it is doubtful that is was
ever really needed.