[io] generated serialization headers from Doxygen
Order is a bit different but respects the same priorities and the same
classes are verified to be included (but fixed for recent
modifications).

TODO: One exception is that DynamicalSystemsGraph and
InteractionsGraph are not seen as derived from _DynamicalSystemsGraph
and _InteractionsGraph, since these latter are typedefs of template
classes, and therefore not included in the class list.