[io serialization] Generated headers; members in alphabetical order.
Previous manual edits to the generated headers are here replaced by
automatic generation from Doxygen XML output.  The main difference is
that members are kept in alphabetical order.  Although a bit
arbitrary, IMHO this is useful to maintain some determinism in where
they should go, making it easier in the future to "diff" generated
files.