[mechanics proposed] The BulletBroadphase no longer needs the model.
It is now an "interaction manager" and gets attached to the simulation
itself.  The simulation gives it an opportunity to update links on
each time step before computing body dynamics.  It uses this
opportunity to keep the collision world synchronised.

TODO: Inform the interaction manager when dynamical systems are removed.