[mechanics proposed] Complete re-factor of BulletBroadphase
Instead of a bunch of messy maps in all directions, now uses a set of
records for each shape type associating the BodyDS with a Siconos
shape, a Bullet shape, a Bullet collision object, and an offset.

Much less use of visitors and single-purpose classes, although a
visitor is still used by the InteractionManager to provide DS position
updates. No more "dirty" lists either, and buildGraph is replaced by
a more descriptive (and internal) createCollisionObject.

Shape sharing is now supported.