[io] SerializerGeneration: Write only which tool was used to generate, avoid full paths
We write the entire generation command to the file but this includes
full paths to where siconos is installed and built on the local
machine, so every time we generate we get changes to the file
depending on where it was done.

This patch makes it note which tool was used, but avoids to write the
full paths which helps with controlling changes to this file.

The full command can be seen by compiling WITH_GENERATION=ON and executing,

    make VERBOSE=1 SerializersGeneration