[kernel] Avoid bad_index exception in nslaw matrix for contact_groups without nslaws.
It should be possible to disable contact in a simulation by removing
the NSLaws, but currently this causes a ublas exception in
symmetric_matrix.  This patch checks the indexes before looking up the
nslaw and returns a null pointer if they exceed the matrix size.

The alternative is to catch the bad_index exception, but this causes
an error message to be printed even in release build.