[mechanics] Set btPersistentManifold's breakingContactThreshold very low.
Fixes contact points being replaced too easily, leading to objects
falling though the floor.

Setting it in the btPersistentManifold instead of globally with
gContactBreakingThreshold decouples this change from the use of
gBreakingContactThreshold in btConvexConvexAlgorithm.cpp for
determining perturbation angle.

This patch also sets the default perturbation iterations back to 3,
since that was the wrong fix.

Closes #106.