[kernel] Move all DS insertion and initialization to NSDS::insertDynamicalSystem().
NSDS::insertDynamicalSystem also calls setOSI() and the OSI's
initializeDynamicalSystem().  Put some common OSI work vector
initialization code in OneStepIntegrator, called from all OSI deriving
classes in initializeDynamicalSystem(), so that work vectors and
matrices are initialized when a DS is added to the graph dynamically.
If the OSI is not yet initialized, this will be deferred until
Simulation::initialize() as before.