[mechanics] fix PrismaticJointR in case that q1/q2 are non-zero and q1cq2 is orthogonal
The real change is the definition of cq2q1 instead of q1cq2, the rest
is just variable renaming.

The reason is that q2*(q10 * q20^-1) != q2*(q20^-1 * q10), and the LHS
gives problems in the case that initial orientations are orthogonal,
because q1 != q2*q1*q2^-1, but q1 == q2*q2^-1*q1.