[mechanics] better CylinderJointR, two pivots with prismatic freedom along an axis.
Previous attempt worked but did not use the point P in the two-DS
case, which was confusing.  This implementation is more like
PivotJointR projected onto a plane, with freedom along the orthogonal
axis.  The two objects can have different radii from the axis, so it
is like two stacked hinges on a rod.  If the axis goes through one of
the objects (e.g. P=0,0,0), it is a piston joint.