[mechanics] fix a few bugs in contact point management, remove params from BulletR.
The "outsideMargin" feature is just not as simple as subtracting from
the distance and moving the contact points along the contact normal.
The BulletR class is thus not the right place to store the scaling and
offset information.  The contact relation is just passed the "correct"
contact point locations, which must be adjusted prior if necessary.

The contact point locations and contact normal are stored relative to
the DS positions, and the world-frame coordinates calculated in
computeh(), so that they move with respect to the Newton loop.