[mechanics_io] rename _broadphase to _interman (param name, interaction_manager)
The name broadphase is not really appropriate since it performs
narrow-phase as well (generates contact points, not just collision
pairs), even though Siconos performs an even finer culling of these
points to determine indexSet1.

Let's call it what it is, an object that manages dynamically the
Interaction objects in the NSDS.

This commit also changes some comments to indicate that objects are
now added to and removed from the NSDS, not the "broadphase".