[kernel] expose bug about workVectors not being initialized
The changes to conditionnaly compile ccertain test when there is a
fortran compiler were buggy. Hence, those tests where never compiled,
nor run. This commit fixes that and fix the compilation issues.
There are some issues with recent changes, resulting in a SEGFAULT.
An assert is now showing the issue.

@vacary @radarsat1 any ideas how to fix it?